CSP-791 Moved GetMyApprenticeshipRequest under correct folder
