+++ src/ApprenticeAan/SFA.DAS.ApprenticeAan.Application.UnitTests/MyApprenticeships/Queries/GetMyApprenticeships/GetMyApprenticeshipRequestTests.cs
@@ -1,5 +1,5 @@
 ﻿using AutoFixture.NUnit3;
-using SFA.DAS.ApprenticeAan.Application.InnerApi.Standards.Requests;
+using SFA.DAS.ApprenticeAan.Application.InnerApi.MyApprenticeships;
 
 namespace SFA.DAS.ApprenticeAan.Application.UnitTests.MyApprenticeships.Queries.GetMyApprenticeships;
