+++ src/ApprenticeAan/SFA.DAS.ApprenticeAan.Application/MyApprenticeships/Queries/GetMyApprenticeship/GetMyApprenticeshipQueryHandler.cs
@@ -1,5 +1,6 @@
 ﻿using System.Net;
 using MediatR;
+using SFA.DAS.ApprenticeAan.Application.InnerApi.MyApprenticeships;
 using SFA.DAS.ApprenticeAan.Application.InnerApi.Standards.Requests;
 using SFA.DAS.SharedOuterApi.Configuration;
 using SFA.DAS.SharedOuterApi.Interfaces;
